+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated Grafana Plugin SDK to 0.4.2.
+
 ## [1.0.1] - 2022-08-12
 
 ### Changed